requirements.txt
absl-py==0.11.0
appdirs==1.4.3
asgiref==3.3.1
astunparse==1.6.3
attr==0.3.1
attrs==20.3.0
boto3==1.16.59
botocore==1.19.59
CacheControl==0.12.6
cachetools==4.2.1
certifi==2019.11.28
cffi==1.14.4
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
coremltools==4.0b3
cryptography==3.3.1
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
dj-database-url==0.5.0
Django==3.1.3
django-allauth==0.44.0
django-ckeditor==6.0.0
django-cors-headers==3.6.0
django-js-asset==1.2.2
django-rest-auth==0.9.5
django-rest-knox==4.1.0
djangorestframework==3.12.2
dynaconf==3.1.2
flatbuffers==1.12
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
gunicorn==20.0.4
h5py==2.10.0
html5lib==1.0.1
hvac==0.10.6
idna==2.8
ipaddr==2.2.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==1.0.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
knox==0.1.11
llvmlite==0.33.0
lockfile==0.12.2
loguru==0.5.3
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.4
mpmath==1.1.0
msgpack==0.6.2
numba==0.50.1
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.3
pandas==1.2.1
pep517==0.8.2
Pillow==8.0.1
prettytable==0.7.2
progress==1.5
protobuf==3.14.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.6
python-dateutil==2.8.1
python3-openid==3.2.0
pytoml==0.1.21
pytz==2020.4
requests==2.22.0
requests-oauthlib==1.3.0
resampy==0.2.1
retrying==1.3.3
rsa==4.7
s3transfer==0.3.4
scikit-learn==0.24.1
scipy==1.6.0
shortuuid==1.0.1
six==1.15.0
sklearn==0.0
sqlparse==0.4.1
sympy==1.7.1
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
threadpoolctl==2.1.0
toml==0.10.2
tqdm==4.56.0
turicreate==6.4.1
typing-extensions==3.7.4.3
urllib3==1.25.8
validators==0.18.2
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.12.1
wheel==0.35
